+++ changelog.md
@@ -1,7 +1,19 @@
 # Changelog
 
-## 0.0.9
+## 0.0.9 Unclear Passes
 
+- **!IMPORTANT!** refactors `GraphicalPass`:
+  - gets rid of 2 generic parameters (descriptor and present-pass markers)
+  - forces the user to create framebuffers
+- **!IMPORTANT!** adds `GraphicalPassBuilder::add_attachment*()` functionality
+  - allows to set user-defined load and store operators
+  - unify workflow for using swapchain and user-declared image resources
+- **!IMPORTANT CHANGE!** `GraphicalPassBuilder::build_present_pass()` is now just `GraphicalPassBuilder::build()`
+- **!IMPORTANT CHANGE!** `Frame::begin_pass()` now takes a framebuffer argument
+- adds `GraphicalPass::begin_framebuffer()` method to start building a framebuffer *(see examples for help)*
+- adds `Frame::get_swapchain_image()` and `Frame::get_swapchain_depth()` accessors to current images for building framebuffers
+- adds `Device::with_depth_format()` constructor for specifying custom swapchain depth format
+- minorly improves documentation
 - updates project readme
 
 ## 0.0.8 Mutated Immutable Docs